feat: Added multiple analytics items for todo including: Consistency metrics for items Weekly Item completion card This month completion card
